fix(external-review-gate): close masking boundary
